memproteksi api tag menggunakan CASL

+++ app/tag/controller.js
@@ -1,8 +1,16 @@
 const Tag = require("./model");
+const { policyFor } = require("../policy");
 
 //add POST
 async function store(req, res, next) {
   try {
+    let policy = policyFor(req.user);
+    if (!policy.can("create", "Tag")) {
+      return res.json({
+        error: 1,
+        message: `Anda tidak memiliki akses untuk menambahkan Tag`,
+      });
+    }
     let payload = req.body;
     let tag = new Tag(payload);
     await tag.save();
@@ -22,6 +30,13 @@ async function store(req, res, next) {
 //update PUT
 async function update(req, res, next) {
   try {
+    let policy = policyFor(req.user);
+    if (!policy.can("update", "Tag")) {
+      return res.json({
+        error: 1,
+        message: `Anda tidak memiliki akses untuk mengupdate Tag`,
+      });
+    }
     let payload = req.body;
     let tag = await Tag.findOneAndUpdate({ _id: req.params.id }, payload, {
       new: true,
@@ -43,6 +58,13 @@ async function update(req, res, next) {
 //destroy DELTED
 async function destroy(req, res, next) {
   try {
+    let policy = policyFor(req.user);
+    if (!policy.can("delete", "Tag")) {
+      return res.json({
+        error: 1,
+        message: `Anda tidak memiliki akses untuk menghapus Tag`,
+      });
+    }
     let tag = await Tag.findOneAndDelete({ _id: req.params.id });
     return res.json(tag);
   } catch (error) {
